refactor: 简化代码写法 调整注释
