+++ src/pages/lessonstable/_components/lesson-popover/index.module.scss
